@@ -1,6 +1,6 @@
 name: veteranam
 description: veteran
-version: 0.0.7+65
+version: 0.0.7+66
 publish_to: none
 
 environment:
@@ -10,63 +10,63 @@ dependencies:
   animated_text_kit: ^4.2.2
   app_tracking_transparency: ^2.0.6
   bloc: ^8.1.4
-  cached_network_image: ^3.4.0
+  cached_network_image: ^3.4.1
   cloud_firestore: ^5.2.1
   collection: ^1.18.0
-  connectivity_plus: ^6.0.4
+  connectivity_plus: ^6.1.1
   dartz: ^0.10.1
-  device_info_plus: ^10.1.2
+  device_info_plus: ^11.2.0
   dio: ^5.7.0
   easy_debounce: ^2.0.3
   email_validator: ^3.0.0
-  equatable: ^2.0.5
+  equatable: ^2.0.7
   feedback: ^3.1.0
-  file_picker: ^8.0.7
-  firebase_analytics: ^11.3.0
-  firebase_app_check: ^0.3.0+4
-  firebase_auth: ^5.2.0
-  firebase_core: ^3.4.0
-  firebase_crashlytics: ^4.1.0
-  firebase_messaging: ^15.1.0
-  firebase_performance: ^0.10.0+4
-  firebase_remote_config: ^5.1.0
-  firebase_storage: ^12.2.0
+  file_picker: ^8.1.6
+  firebase_analytics: ^11.3.6
+  firebase_app_check: ^0.3.1
+  firebase_auth: ^5.3.4
+  firebase_core: ^3.8.1
+  firebase_crashlytics: ^4.2.0
+  firebase_messaging: ^15.1.6
+  firebase_performance: ^0.10.0
+  firebase_remote_config: ^5.2.0
+  firebase_storage: ^12.3.7
   flutter:
     sdk: flutter
   flutter_bloc: ^8.1.6
-  flutter_facebook_auth: ^7.1.0
+  flutter_facebook_auth: ^7.1.1
   flutter_localizations:
     sdk: flutter
-  flutter_markdown: ^0.7.3
+  flutter_markdown: ^0.7.4
   flutter_rating_bar: ^4.0.1
   flutter_secure_storage: ^9.2.2
-  formz: ^0.7.0
+  formz: ^0.8.0
   freezed: ^2.5.7
   freezed_annotation: ^2.4.4
-  get_it: ^7.7.0 # Used for dependency injection
-  go_router: ^14.2.7 # Decoupled routing  package which is a must for any app
-  google_sign_in: ^6.2.1
+  get_it: ^8.0.3 # Used for dependency injection
+  go_router: ^14.6.2 # Decoupled routing  package which is a must for any app
+  google_sign_in: ^6.2.2
   image_picker: ^1.1.2
-  in_app_review: ^2.0.9
-  injectable: ^2.4.4 # Used for dependency injection
+  in_app_review: ^2.0.10
+  injectable: ^2.5.0 # Used for dependency injection
   intl: ^0.19.0
-  json_serializable: ^6.8.0
-  material_symbols_icons: ^4.2784.0
-  package_info_plus: ^8.0.1
+  json_serializable: ^6.9.0
+  material_symbols_icons: ^4.2801.0
+  package_info_plus: ^8.1.2
   # path_drawing: ^1.0.1
   # pdf: ^3.11.0
   # printing: ^5.13.1
-  path_provider: ^2.1.4
-  sentry_flutter: ^8.8.0
-  share_plus: ^10.0.0
+  path_provider: ^2.1.5
+  sentry_flutter: ^8.11.1
+  share_plus: ^10.1.3
   skeletonizer: ^1.4.2
-  url_launcher: ^6.3.0
-#  web: 0.5.1 # Wait when share plus support version 1.0.0
+  url_launcher: ^6.3.1
+  # web: 0.5.1 # Wait when share plus support version 1.0.0
 
 dependency_overrides:
   flutter_secure_storage_web: ^2.0.0-beta.1
   meta: ^1.15.0
-  # web: ^1.1.0
+  web: ^1.1.0
 
 #   vector_graphics_compiler:
 #     git:
