@@ -1,66 +1,35 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:veteranam/components/home/bloc/home_watcher_bloc.dart';
-import 'package:veteranam/components/home/home.dart';
 import 'package:veteranam/shared/shared_flutter.dart';
 
-class FaqSectionWidget extends StatelessWidget {
-  const FaqSectionWidget({
-    required this.isDesk,
-    required this.padding,
-    required this.isTablet,
+class FAQSectionDeskWidget extends StatelessWidget {
+  const FAQSectionDeskWidget({
     super.key,
   });
-  final bool isDesk;
-  final bool isTablet;
-  final EdgeInsets padding;
 
   @override
   Widget build(BuildContext context) {
-    if (isDesk) {
-      return SliverPadding(
-        padding: padding,
-        sliver: SliverToBoxAdapter(
-          child: _FAQSectionDeskWidget(
-            isDesk: isDesk,
-            isTablet: isTablet,
-          ),
-        ),
-      );
-    } else {
-      return _FaqSectionMobWidget(
-        isTablet: isTablet,
-        padding: padding,
-      );
-    }
-  }
-}
-
-class _FAQSectionDeskWidget extends StatelessWidget {
-  const _FAQSectionDeskWidget({
-    required this.isDesk,
-    required this.isTablet,
-  });
-  final bool isDesk;
-  final bool isTablet;
-
-  @override
-  Widget build(BuildContext context) {
-    return const Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Expanded(
+    return const SliverCrossAxisGroup(
+      slivers: [
+        SliverCrossAxisExpanded(
           flex: 2,
-          child: _GetFAQSection(
-            isDesk: true,
+          sliver: SliverPadding(
+            padding: EdgeInsets.only(top: KPadding.kPaddingSize24),
+            sliver: SliverToBoxAdapter(
+              child: _GetFAQSection(
+                isDesk: true,
+              ),
+            ),
           ),
         ),
-        KSizedBox.kWidthSizedBox8,
-        Expanded(
+        SliverToBoxAdapter(
+          child: KSizedBox.kWidthSizedBox8,
+        ),
+        SliverCrossAxisExpanded(
           flex: 3,
-          child: QuestionWidgetList(
+          sliver: _FaqListWidget(
             isDesk: true,
           ),
         ),
@@ -69,72 +38,81 @@ class _FAQSectionDeskWidget extends StatelessWidget {
   }
 }
 
-class _FaqSectionMobWidget extends SingleChildRenderObjectWidget {
-  _FaqSectionMobWidget({
-    required this.isTablet,
-    required this.padding,
-  }) : super(
-          child: _FaqSectionMobWidgetImplementation(
-            isTablet: isTablet,
-          ),
-        );
-
-  final bool isTablet;
-  final EdgeInsets padding;
+class FaqSectionMobWidget extends StatelessWidget {
+  const FaqSectionMobWidget({super.key});
 
   @override
-  RenderSliverPadding createRenderObject(BuildContext context) {
-    return RenderSliverPadding(
-      padding: padding,
-      textDirection: Directionality.of(context),
+  Widget build(BuildContext context) {
+    return const SliverMainAxisGroup(
+      slivers: [
+        SliverToBoxAdapter(
+          child: _GetFAQSection(
+            isDesk: false,
+          ),
+        ),
+        _FaqListWidget(
+          isDesk: false,
+        ),
+      ],
     );
   }
 }
 
-class _FaqSectionMobWidgetImplementation extends StatelessWidget {
-  const _FaqSectionMobWidgetImplementation({required this.isTablet});
-  final bool isTablet;
+class _FaqListWidget extends StatelessWidget {
+  const _FaqListWidget({required this.isDesk});
+  final bool isDesk;
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<HomeWatcherBloc, HomeWatcherState>(
-      buildWhen: (previous, current) =>
-          previous.loadingStatus != current.loadingStatus,
-      builder: (context, state) {
-        return SliverMainAxisGroup(
-          slivers: [
-            const SliverToBoxAdapter(
-              child: _GetFAQSection(
-                isDesk: false,
-              ),
-            ),
-            BlocBuilder<LanguageCubit, Language>(
-              builder: (context, language) {
-                return SliverList.builder(
-                  itemBuilder: (context, index) => Padding(
-                    key: ValueKey(KMockText.questionModel.id),
-                    padding: const EdgeInsets.only(
-                      top: KPadding.kPaddingSize24,
-                    ),
-                    child: QuestionWidgetItem(
-                      state: state,
-                      index: index,
-                      isDesk: false,
-                    ),
+    return BlocBuilder<LanguageCubit, Language>(
+      builder: (context, language) {
+        return BlocBuilder<HomeWatcherBloc, HomeWatcherState>(
+          builder: (context, state) {
+            if (state.loadingStatus.isLoading) {
+              return SliverPrototypeExtentList.builder(
+                prototypeItem: mockQuestionWidget,
+                itemBuilder: (context, index) => mockQuestionWidget,
+                addAutomaticKeepAlives: false,
+                addRepaintBoundaries: false,
+                itemCount: KDimensions.shimmerQuestionItems,
+              );
+            } else {
+              return SliverList.builder(
+                itemBuilder: (context, index) => Padding(
+                  key: ValueKey(KMockText.questionModel.id),
+                  padding: const EdgeInsets.only(
+                    top: KPadding.kPaddingSize24,
                   ),
-                  addAutomaticKeepAlives: false,
-                  addRepaintBoundaries: false,
-                  itemCount: (state.loadingStatus == LoadingStatusHome.loaded
-                      ? state.questionModelItems.length
-                      : KDimensions.shimmerQuestionItems),
-                );
-              },
-            ),
-          ],
+                  child: QuestionWidget(
+                    key: KWidgetkeys.screen.home.faq,
+                    questionModel: state.questionModelItems.elementAt(index),
+                    isDesk: isDesk,
+                  ),
+                ),
+                addAutomaticKeepAlives: false,
+                addRepaintBoundaries: false,
+                itemCount: state.questionModelItems.length,
+              );
+            }
+          },
         );
       },
     );
   }
+
+  Widget get mockQuestionWidget => Padding(
+        key: ValueKey(KMockText.questionModel.id),
+        padding: const EdgeInsets.only(
+          top: KPadding.kPaddingSize24,
+        ),
+        child: SkeletonizerWidget(
+          isLoading: true,
+          child: QuestionWidget(
+            questionModel: KMockText.questionModel,
+            isDesk: isDesk,
+          ),
+        ),
+      );
 }
 
 class _GetFAQSection extends StatelessWidget {
