@@ -23,8 +23,8 @@ class LoginBodyWidget extends StatelessWidget {
           ),
           isForm: true,
           pageName: context.l10n.login,
-          showMobNawbarBackButton: true,
-          showMobBottomNavigation: false,
+          showMobBottomNavigation: true,
+          bottomBarIndex: 3,
           mainChildWidgetsFunction: ({required isDesk, required isTablet}) => [
             if (!isDesk) KSizedBox.kHeightSizedBox24,
             if (Config.isWeb)
