@@ -13,9 +13,14 @@ class ImagesConverter
 
   @override
   List<ImageModel>? fromJson(List<dynamic>? json) {
-    final list = json
-        ?.map((e) => ImageModel?.fromJson(e as Map<String, dynamic>))
-        .toList();
+    final list = json == null
+        ? null
+        : List.generate(
+            json.length,
+            (index) => ImageModel.fromJson(
+              json.elementAt(index) as Map<String, dynamic>,
+            ),
+          );
     if (list?.isEmpty ?? true) {
       return null;
     } else {
