@@ -92,7 +92,7 @@ class SomeFailure with _$SomeFailure {
   }) = FailureUnauthorized;
 
   factory SomeFailure.notFound({
-    Object? error,
+    required Object? error,
     StackTrace? stack,
     ErrorLevelEnum? errorLevel,
     String? tag,
@@ -105,7 +105,7 @@ class SomeFailure with _$SomeFailure {
   }) = FailureNotFound;
 
   factory SomeFailure.duplicate({
-    Object? error,
+    required Object? error,
     StackTrace? stack,
     ErrorLevelEnum? errorLevel,
     String? tag,
@@ -158,7 +158,7 @@ class SomeFailure with _$SomeFailure {
   }) = FailureShare;
 
   factory SomeFailure.shareUnavailable({
-    Object? error,
+    required Object? error,
     StackTrace? stack,
     ErrorLevelEnum? errorLevel,
     String? tag,
@@ -197,7 +197,7 @@ class SomeFailure with _$SomeFailure {
   }) = FailureCopy;
 
   factory SomeFailure.wrongVerifyCode({
-    Object? error,
+    required Object? error,
     StackTrace? stack,
     ErrorLevelEnum? errorLevel,
     String? tag,
@@ -210,7 +210,7 @@ class SomeFailure with _$SomeFailure {
   }) = FailureWrongVerifyCode;
 
   factory SomeFailure.filter({
-    Object? error,
+    required Object? error,
     StackTrace? stack,
     ErrorLevelEnum? errorLevel,
     String? tag,
@@ -223,7 +223,7 @@ class SomeFailure with _$SomeFailure {
   }) = FailureFilter;
 
   factory SomeFailure.browserNotSupportPopupDialog({
-    Object? error,
+    required Object? error,
     StackTrace? stack,
     ErrorLevelEnum? errorLevel,
     String? tag,
@@ -234,6 +234,19 @@ class SomeFailure with _$SomeFailure {
     UserSetting? userSetting,
     String? data,
   }) = FailureBrowserNotSupportPopupDialog;
+
+  factory SomeFailure.emailInvalidFormat({
+    required Object? error,
+    StackTrace? stack,
+    ErrorLevelEnum? errorLevel,
+    String? tag,
+    String? tagKey,
+    String? tag2,
+    String? tag2Key,
+    User? user,
+    UserSetting? userSetting,
+    String? data,
+  }) = FailureEmailInvalidFormat;
 
   final FailureRepository failureRepository = GetIt.I.get<FailureRepository>();
 }
