@@ -233,7 +233,6 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
         SomeFailure.filter(
           error: e,
           stack: stack,
-          data: unmodifiedDiscountModelItems.toString(),
           tag: 'getFilterList',
           tagKey: 'Discount Filter ${ErrorText.repositoryKey}',
         ),
@@ -341,7 +340,7 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
         SomeFailure.filter(
           error: e,
           stack: stack,
-          data: '$isEnglish |$unmodifiedDiscountModelItems',
+          data: 'isEnglish: $isEnglish',
           tag: 'getFilterValuesFromDiscountItems',
           tagKey: 'Discount Filter ${ErrorText.repositoryKey}',
           tag2Key: ErrorText.callFrom,
@@ -370,10 +369,17 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
             TranslateModel(uk: valueUK),
           );
       if (!activityFilter.containsKey(valueUK)) {
+        // One Item Code Start:
+        for (final key in activityFilter.keys) {
+          filter[key] = filter[key]!.copyWith(isSelected: false);
+        }
+        activityFilter.clear();
+        // One Item Code End:
         activityFilter[valueUK] = filterItem;
       } else {
         activityFilter.remove(valueUK);
       }
+
       filter[valueUK] = filterItem;
       // Add New Filter Item To activity List and Change Is Selected For Item
       // With
@@ -506,8 +512,8 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
         SomeFailure.filter(
           error: e,
           stack: stack,
-          data: '$valueUK | $filter | $activityFilter '
-              '| $filterEnum $unmodifiedDiscountModelItems',
+          data: 'valueUK: $valueUK | activityFilter: $activityFilter '
+              '| filterEnum: $filterEnum',
           tag: '_addFilterItem',
           tagKey: 'Discount Filter ${ErrorText.repositoryKey}',
           tag2Key: ErrorText.callFrom,
@@ -536,7 +542,7 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
         SomeFailure.filter(
           error: e,
           stack: stack,
-          data: '$activityMap | $itemsMap',
+          data: 'activityMap: $activityMap',
           tag: '_addActivityMapToItemsMap',
           tagKey: 'Discount Filter ${ErrorText.repositoryKey}',
           tag2Key: ErrorText.callFrom,
@@ -610,7 +616,7 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
       SomeFailure.filter(
         error: e,
         stack: stack,
-        data: '$values | $activityFilter',
+        data: 'activityFilter: $activityFilter',
         tag: '_activityListContainAnyValues',
         tagKey: 'Discount Filter ${ErrorText.repositoryKey}',
         tag2Key: ErrorText.callFrom,
@@ -667,7 +673,7 @@ class DiscountFilterRepository implements IDiscountFilterRepository {
         SomeFailure.filter(
           error: e,
           stack: stack,
-          data: '$list | $activityMap',
+          data: 'activityMap: $activityMap',
           tag: '_getFilterFromTranslateModel',
           tagKey: 'Discount Filter ${ErrorText.repositoryKey}',
           tag2Key: ErrorText.callFrom,
