@@ -11,6 +11,7 @@ class CancelChipWidget extends StatelessWidget {
     this.style,
     this.textStyle,
     this.padding,
+    // this.width,
   });
   final Key widgetKey;
   final bool isDesk;
@@ -19,26 +20,46 @@ class CancelChipWidget extends StatelessWidget {
   final ButtonStyle? style;
   final TextStyle? textStyle;
   final EdgeInsets? padding;
+  // final double? width;
 
   @override
   Widget build(BuildContext context) {
-    return TextButton.icon(
+    return TextButton(
       key: widgetKey,
       style: (style ?? KButtonStyles.advancedFilterButtonStyle).copyWith(
         padding: padding == null ? null : WidgetStatePropertyAll(padding),
+        // maximumSize: width == null
+        //     ? null
+        //     : WidgetStatePropertyAll(Size(width!, double.infinity)),
+        // minimumSize:
+        //     width == null ? null : WidgetStatePropertyAll(Size(width!, 0)),
       ),
-      icon: KIcon.close,
       clipBehavior: Clip.hardEdge,
-      label: Text(
+      onPressed: onPressed,
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          KIcon.close,
+          KSizedBox.kWidthSizedBox8,
+          // if (width == null)
+          Flexible(child: textWidget),
+          // else
+          //   Expanded(
+          //     child: textWidget,
+          //   ),
+        ],
+      ),
+    );
+  }
+
+  Widget get textWidget => Text(
         labelText,
         textAlign: TextAlign.center,
         maxLines: 1,
+        overflow: TextOverflow.ellipsis,
         style: textStyle ??
             (isDesk
                 ? AppTextStyle.materialThemeBodyLarge
                 : AppTextStyle.materialThemeBodyMedium),
-      ),
-      onPressed: onPressed,
-    );
-  }
+      );
 }
