@@ -1,155 +1,156 @@
-// ignore_for_file: library_private_types_in_public_api
+// // ignore_for_file: library_private_types_in_public_api
 
-import 'dart:math';
-import 'package:flutter/rendering.dart';
+// import 'dart:math';
+// import 'package:flutter/rendering.dart';
 
-// class RowSliver extends MultiChildRenderObjectWidget {
-//   // Constructor for the RowSliver widget
-//   RowSliver({
-//     required this.left,
-//     required this.right,
-//     required this.leftWidthPercent,
-//     super.key,
-//   }) : super(children: [left, right]);
+// // class RowSliver extends MultiChildRenderObjectWidget {
+// //   // Constructor for the RowSliver widget
+// //   RowSliver({
+// //     required this.left,
+// //     required this.right,
+// //     required this.leftWidthPercent,
+// //     super.key,
+// //   }) : super(children: [left, right]);
 
-//   final Widget left; // Left widget to display
-//   final Widget right; // Right widget to display
-//   final double
-//       leftWidthPercent; // Percentage of width allocated to the left widget
+// //   final Widget left; // Left widget to display
+// //   final Widget right; // Right widget to display
+// //   final double
+// //       leftWidthPercent; // Percentage of width allocated to the left widget
 
-//   // Creates the render object for this widget
-//   @override
-//   RenderRowSliver createRenderObject(BuildContext context) {
-//     return RenderRowSliver(leftWidthPercent: leftWidthPercent);
-//   }
+// //   // Creates the render object for this widget
+// //   @override
+// //   RenderRowSliver createRenderObject(BuildContext context) {
+// //     return RenderRowSliver(leftWidthPercent: leftWidthPercent);
+// //   }
 
-//   @override
-//   void updateRenderObject(
-//     BuildContext context,
-//     RenderRowSliver renderObject,
-//   ) {}
+// //   @override
+// //   void updateRenderObject(
+// //     BuildContext context,
+// //     RenderRowSliver renderObject,
+// //   ) {}
+// // }
+
+// // Extension to get custom parent data for RowSliver
+// extension _RowSliderParentDataExt on RenderSliver {
+//   _RowSliverParentData get rowSliver => parentData! as _RowSliverParentData;
 // }
 
-// Extension to get custom parent data for RowSliver
-extension _RowSliderParentDataExt on RenderSliver {
-  _RowSliverParentData get rowSliver => parentData! as _RowSliverParentData;
-}
+// // Custom parent data class to hold additional data for each sliver child
+// class _RowSliverParentData extends SliverPhysicalParentData
+//     with ContainerParentDataMixin<RenderSliver> {}
 
-// Custom parent data class to hold additional data for each sliver child
-class _RowSliverParentData extends SliverPhysicalParentData
-    with ContainerParentDataMixin<RenderSliver> {}
+// class RenderRowSliver extends RenderSliver
+//     with ContainerRenderObjectMixin<RenderSliver, _RowSliverParentData> {
+//   // Constructor to initialize the width percentage for the left widget
+//   RenderRowSliver({required double leftWidthPercent})
+//       : _leftWidthPercent = leftWidthPercent;
 
-class RenderRowSliver extends RenderSliver
-    with ContainerRenderObjectMixin<RenderSliver, _RowSliverParentData> {
-  // Constructor to initialize the width percentage for the left widget
-  RenderRowSliver({required double leftWidthPercent})
-      : _leftWidthPercent = leftWidthPercent;
+//   final double _leftWidthPercent;
 
-  final double _leftWidthPercent;
+//   // Get references to the left and right children
+//   RenderSliver get left => firstChild!;
+//   RenderSliver get right => lastChild!;
 
-  // Get references to the left and right children
-  RenderSliver get left => firstChild!;
-  RenderSliver get right => lastChild!;
+//   // Sets up the parent data for the sliver children
+//   @override
+//   void setupParentData(RenderSliver child) {
+//     if (child.parentData is! _RowSliverParentData) {
+//       child.parentData = _RowSliverParentData();
+//     }
+//   }
 
-  // Sets up the parent data for the sliver children
-  @override
-  void setupParentData(RenderSliver child) {
-    if (child.parentData is! _RowSliverParentData) {
-      child.parentData = _RowSliverParentData();
-    }
-  }
+//   // Perform layout calculations for the sliver children
+//   @override
+//   void performLayout() {
+//     if (firstChild == null) {
+//       geometry = SliverGeometry.zero;
+//       return;
+//     }
 
-  // Perform layout calculations for the sliver children
-  @override
-  void performLayout() {
-    if (firstChild == null) {
-      geometry = SliverGeometry.zero;
-      return;
-    }
+//     // Calculate the size for the left child based on the width percentage
+//     final leftSize = constraints.crossAxisExtent * _leftWidthPercent;
+//     final remainingWidth = constraints.crossAxisExtent - leftSize;
 
-    // Calculate the size for the left child based on the width percentage
-    final leftSize = constraints.crossAxisExtent * _leftWidthPercent;
-    final remainingWidth = constraints.crossAxisExtent - leftSize;
+//     // Layout both children with the respective sizes
+//     left.layout(
+//       constraints.copyWith(crossAxisExtent: leftSize),
+//       parentUsesSize: true,
+//     );
+//     right.layout(
+//       constraints.copyWith(crossAxisExtent: remainingWidth),
+//       parentUsesSize: true,
+//     );
 
-    // Layout both children with the respective sizes
-    left.layout(
-      constraints.copyWith(crossAxisExtent: leftSize),
-      parentUsesSize: true,
-    );
-    right.layout(
-      constraints.copyWith(crossAxisExtent: remainingWidth),
-      parentUsesSize: true,
-    );
+//     // Calculate the maximum scroll extent based on the children
+//     final maxScrollExtent =
+//         max(left.geometry!.scrollExtent, right.geometry!.scrollExtent);
+//     final paintExtent = min(constraints.remainingPaintExtent,
+// maxScrollExtent);
 
-    // Calculate the maximum scroll extent based on the children
-    final maxScrollExtent =
-        max(left.geometry!.scrollExtent, right.geometry!.scrollExtent);
-    final paintExtent = min(constraints.remainingPaintExtent, maxScrollExtent);
+//     // Set the sliver geometry based on the calculations
+//     geometry = SliverGeometry(
+//       scrollExtent: maxScrollExtent,
+//       paintExtent: paintExtent,
+//       layoutExtent: paintExtent,
+//       maxPaintExtent: maxScrollExtent,
+//       cacheExtent: maxScrollExtent,
+//     );
 
-    // Set the sliver geometry based on the calculations
-    geometry = SliverGeometry(
-      scrollExtent: maxScrollExtent,
-      paintExtent: paintExtent,
-      layoutExtent: paintExtent,
-      maxPaintExtent: maxScrollExtent,
-      cacheExtent: maxScrollExtent,
-    );
+//     (left.parentData! as _RowSliverParentData).paintOffset = Offset.zero;
 
-    (left.parentData! as _RowSliverParentData).paintOffset = Offset.zero;
+//     (right.parentData! as _RowSliverParentData).paintOffset =
+//         Offset(leftSize, 0);
+//   }
 
-    (right.parentData! as _RowSliverParentData).paintOffset =
-        Offset(leftSize, 0);
-  }
+//   // Paint the children to the canvas
+//   @override
+//   void paint(PaintingContext context, Offset offset) {
+//     if (!geometry!.visible) return;
 
-  // Paint the children to the canvas
-  @override
-  void paint(PaintingContext context, Offset offset) {
-    if (!geometry!.visible) return;
+//     // Paint the left child first, then the right child with a horizontal offset
+//     final leftParentData = left.parentData! as _RowSliverParentData;
+//     final rightParentData = right.parentData! as _RowSliverParentData;
 
-    // Paint the left child first, then the right child with a horizontal offset
-    final leftParentData = left.parentData! as _RowSliverParentData;
-    final rightParentData = right.parentData! as _RowSliverParentData;
+//     context
+//       ..paintChild(left, offset + leftParentData.paintOffset)
+//       ..paintChild(right, offset + rightParentData.paintOffset);
+//   }
 
-    context
-      ..paintChild(left, offset + leftParentData.paintOffset)
-      ..paintChild(right, offset + rightParentData.paintOffset);
-  }
+//   // Check if the touch event hit any of the children
+//   @override
+//   bool hitTestChildren(
+//     SliverHitTestResult result, {
+//     required double mainAxisPosition,
+//     required double crossAxisPosition,
+//   }) {
+//     final children = [left, right]; // Check both children
+//     for (final child in children) {
+//       if (child.geometry!.visible) {
+//         var adjustedCrossAxisPosition = crossAxisPosition;
 
-  // Check if the touch event hit any of the children
-  @override
-  bool hitTestChildren(
-    SliverHitTestResult result, {
-    required double mainAxisPosition,
-    required double crossAxisPosition,
-  }) {
-    final children = [left, right]; // Check both children
-    for (final child in children) {
-      if (child.geometry!.visible) {
-        var adjustedCrossAxisPosition = crossAxisPosition;
+//         // For the right child, adjust the position based on the left
+//         // widget's width
+//         if (child == right) {
+//           adjustedCrossAxisPosition -=
+//               constraints.crossAxisExtent * _leftWidthPercent;
+//         }
 
-        // For the right child, adjust the position based on the left
-        // widget's width
-        if (child == right) {
-          adjustedCrossAxisPosition -=
-              constraints.crossAxisExtent * _leftWidthPercent;
-        }
+//         // Now check if the touch is inside the child
+//         if (child.hitTest(
+//           result,
+//           mainAxisPosition: mainAxisPosition,
+//           crossAxisPosition: adjustedCrossAxisPosition,
+//         )) {
+//           return true;
+//         }
+//       }
+//     }
+//     return false;
+//   }
 
-        // Now check if the touch is inside the child
-        if (child.hitTest(
-          result,
-          mainAxisPosition: mainAxisPosition,
-          crossAxisPosition: adjustedCrossAxisPosition,
-        )) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  // Apply the paint transform for the child slivers
-  @override
-  void applyPaintTransform(RenderSliver child, Matrix4 transform) {
-    child.rowSliver.applyPaintTransform(transform);
-  }
-}
+//   // Apply the paint transform for the child slivers
+//   @override
+//   void applyPaintTransform(RenderSliver child, Matrix4 transform) {
+//     child.rowSliver.applyPaintTransform(transform);
+//   }
+// }
