@@ -25,7 +25,7 @@ import package_info_plus
 import path_provider_foundation
 import sentry_flutter
 import share_plus
-import sqflite
+import sqflite_darwin
 import url_launcher_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
